+++ spec/rubocop/cop/rails/save_bang_spec.rb
@@ -1,11 +1,47 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Rails::SaveBang do
-  subject(:cop) { described_class.new }
+RSpec.describe RuboCop::Cop::Rails::SaveBang, :config do
+  subject(:cop) { described_class.new(config) }
 
   shared_examples 'checks_common_offense' do |method|
     it "when using #{method} with arguments" do
       inspect_source("object.#{method}(name: 'Tom', age: 20)")
+
+      if method == :destroy
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["Use `#{method}!` instead of `#{method}` " \
+                  'if the return value is not checked.'])
+      end
+    end
+
+    it "when using #{method} with variable arguments" do
+      inspect_source("object.#{method}(variable)")
+
+      if method == :destroy
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["Use `#{method}!` instead of `#{method}` " \
+                  'if the return value is not checked.'])
+      end
+    end
+
+    it "when using #{method} with variable star arguments" do
+      inspect_source("object.#{method}(*variable)")
+
+      if method == :destroy
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["Use `#{method}!` instead of `#{method}` " \
+                  'if the return value is not checked.'])
+      end
+    end
+
+    it "when using #{method} with variable star star arguments" do
+      inspect_source("object.#{method}(**variable)")
 
       if method == :destroy
         expect(cop.messages.empty?).to be(true)
@@ -43,11 +79,11 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     end
   end
 
-  shared_examples 'checks_variable_return_use_offense' do |method, pass|
+  shared_examples 'checks_variable_return_use_offense' do |method, update|
     it "when assigning the return value of #{method}" do
       inspect_source("x = object.#{method}\n")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -62,7 +98,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
                           "  obj.name = 'Tom'\n" \
                           'end')
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -75,7 +111,48 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with if" do
       inspect_source("if object.#{method}; something; end")
 
-      if pass
+      if update
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["`#{method}` returns a model which is always truthy."])
+      end
+    end
+
+    it "when using #{method} with if with method chain" do
+      inspect_source(<<-RUBY.strip_indent)
+        if object.tap(&:prepare_for_save).#{method}
+          something
+        end
+      RUBY
+
+      if update
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["`#{method}` returns a model which is always truthy."])
+      end
+    end
+
+    it "when using #{method} with if with block" do
+      inspect_source(<<-RUBY.strip_indent)
+        if object.#{method} { |o| o.name = 'Tom' }
+          something
+        end
+      RUBY
+
+      if update
+        expect(cop.messages.empty?).to be(true)
+      else
+        expect(cop.messages)
+          .to eq(["`#{method}` returns a model which is always truthy."])
+      end
+    end
+
+    it "when using #{method} with referenced block" do
+      inspect_source("if object.#{method}(&:values); something; end")
+
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -86,7 +163,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with negated if" do
       inspect_source("if !object.#{method}; something; end")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -100,7 +177,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
           something
         end
       RUBY
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -111,7 +188,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with oneline if" do
       inspect_source("something if object.#{method}")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -122,12 +199,20 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with oneline if and multiple conditional" do
       inspect_source("something if false || object.#{method}")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
           .to eq(["`#{method}` returns a model which is always truthy."])
       end
+    end
+
+    it "when using #{method} with a bunch of hashes & arrays" do
+      inspect_source(<<-RUBY.strip_indent)
+        return [{ success: object.#{method} }, true]
+      RUBY
+
+      expect(cop.messages.empty?).to be(true)
     end
 
     it "when using #{method} with case statement" do
@@ -140,7 +225,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
         end
       RUBY
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -151,7 +236,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with '&&'" do
       inspect_source("object.#{method} && false")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -162,7 +247,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with 'and'" do
       inspect_source("object.#{method} and false")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -173,18 +258,26 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     it "when using #{method} with '||'" do
       inspect_source("object.#{method} || false")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
           .to eq(["`#{method}` returns a model which is always truthy."])
       end
+    end
+
+    it 'when passing to a method' do
+      expect_no_offenses("handle_save(object.#{method})")
+    end
+
+    it 'when passing to a method as the non-last argument' do
+      expect_no_offenses("handle_save(object.#{method}, true)")
     end
 
     it "when using #{method} with 'or'" do
       inspect_source("object.#{method} or false")
 
-      if pass
+      if update
         expect(cop.messages.empty?).to be(true)
       else
         expect(cop.messages)
@@ -192,10 +285,44 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
       end
     end
 
-    it "when using #{method} as last method call" do
+    it 'when passing to a method as a keyword argument' do
+      expect_no_offenses("handle_save(success: object.#{method})")
+    end
+
+    it 'when assigning as a hash value' do
+      expect_no_offenses("result = { success: object.#{method} }")
+    end
+
+    it 'when using an explicit early return' do
       expect_no_offenses(<<-RUBY.strip_indent)
         def foo
-          object.#{method}
+          return foo.#{method} if do_the_save
+          do_something_else
+        end
+      RUBY
+    end
+
+    it 'when using an explicit final return' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        def foo
+          return foo.#{method}
+        end
+      RUBY
+    end
+
+    it 'when using an explicit early return from a block' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        objects.each do |object|
+          next object.#{method} if do_the_save
+          do_something_else
+        end
+      RUBY
+    end
+
+    it 'when using an explicit final return from a block' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        objects.each do |object|
+          next foo.#{method}
         end
       RUBY
     end
@@ -208,7 +335,7 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
           render json: foo
         end
       RUBY
-      if pass
+      if update
         expect(cop.offenses.empty?).to be(true)
       else
         expect(cop.offenses.size).to eq(1)
@@ -216,9 +343,71 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
     end
   end
 
+  shared_examples 'check_implicit_return' do |method, allow_implicit_return|
+    it "when using #{method} as last method call" do
+      inspect_source(<<-RUBY.strip_indent)
+        def foo
+          object.#{method}
+        end
+      RUBY
+
+      if allow_implicit_return
+        expect(cop.offenses.empty?).to be true
+      else
+        expect(cop.messages)
+          .to match_array(start_with("Use `#{method}!` instead of `#{method}`" \
+                             ' if the return value is not checked.'))
+      end
+    end
+
+    it "when using #{method} as last method call of a block" do
+      inspect_source(<<-RUBY.strip_indent)
+        objects.each do |object|
+          object.#{method}
+        end
+      RUBY
+
+      if allow_implicit_return
+        expect(cop.offenses.empty?).to be true
+      else
+        expect(cop.messages)
+          .to match_array(start_with("Use `#{method}!` instead of `#{method}`" \
+                             ' if the return value is not checked.'))
+      end
+    end
+
+    it "when using #{method} as part of the last line" do
+      inspect_source(<<-RUBY.strip_indent)
+        def whatever
+          [{ success: object.#{method} }, true]
+        end
+      RUBY
+
+      if allow_implicit_return
+        expect(cop.offenses.empty?).to be true
+      else
+        expect(cop.messages)
+          .to match_array(start_with("Use `#{method}!` instead of `#{method}`" \
+                             ' if the return value is not checked.'))
+      end
+    end
+  end
+
   described_class::MODIFY_PERSIST_METHODS.each do |method|
-    it_behaves_like('checks_common_offense', method)
-    it_behaves_like('checks_variable_return_use_offense', method, true)
+    let(:cop_config) { { 'AllowImplicitReturn' => true } }
+
+    context method.to_s do
+      it_behaves_like('checks_common_offense', method)
+      it_behaves_like('checks_variable_return_use_offense', method, true)
+      it_behaves_like('check_implicit_return', method, true)
+
+      context 'with AllowImplicitReturn false' do
+        let(:cop_config) { { 'AllowImplicitReturn' => false } }
+
+        it_behaves_like('checks_variable_return_use_offense', method, true)
+        it_behaves_like('check_implicit_return', method, false)
+      end
+    end
   end
 
   shared_examples 'checks_create_offense' do |method|
@@ -233,12 +422,31 @@ RSpec.describe RuboCop::Cop::Rails::SaveBang do
                           "end\n" \
                           'if x.persisted? then; something; end')
     end
+
+    it "when using persisted? after #{method} called on a chain" do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        x = User.unverified.with_initial_values.#{method}
+        if x.persisted?; something; end
+      RUBY
+    end
   end
 
   described_class::CREATE_PERSIST_METHODS.each do |method|
-    it_behaves_like('checks_common_offense', method)
-    it_behaves_like('checks_variable_return_use_offense', method, false)
-    it_behaves_like('checks_create_offense', method)
+    let(:cop_config) { { 'AllowImplicitReturn' => true } }
+
+    context method.to_s do
+      it_behaves_like('checks_common_offense', method)
+      it_behaves_like('checks_variable_return_use_offense', method, false)
+      it_behaves_like('checks_create_offense', method)
+      it_behaves_like('check_implicit_return', method, true)
+
+      context 'with AllowImplicitReturn false' do
+        let(:cop_config) { { 'AllowImplicitReturn' => false } }
+
+        it_behaves_like('checks_variable_return_use_offense', method, false)
+        it_behaves_like('check_implicit_return', method, false)
+      end
+    end
   end
 
   it 'properly ignores lvasign without right hand side' do